Module 7 Implemenattation/requirements.txt
streamlit
opencv-python
numpy
pytesseract
Pillow
easyocr
firebase-admin
